#13 - request, response 패키지 수정
request, response 클래스가 많아지면서
도메인 별로 패키지를 수정했다.
