+++ web-service/src/main/java/com/ticketcheater/webservice/controller/request/member/MemberSignupRequest.java
@@ -1,4 +1,4 @@
-package com.ticketcheater.webservice.controller.request;
+package com.ticketcheater.webservice.controller.request.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberLoginRequest {
+public class MemberSignupRequest {
     private String name;
     private String password;
+    private String email;
+    private String nickname;
 }
